Set up tile maps and added some environment shapes
I'm a bit confused about how the layers work with everything being a part of the Canvas. Wasn't able to get colliders to work correctly or the camera to track the player like I normally would, so definitely need some insight on that.  Also, I somehow made the ship's sprite and exhaust not visible from my end... whoops haha!